+++ simple-stack/src/main/java/com/zhuinden/simplestack/Bundleable.java
@@ -15,19 +15,18 @@
  */
 package com.zhuinden.simplestack;
 
-import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.view.View;
 
 /**
- * Specifies that the custom view that implements this also places its persisted state into a Bundle.
+ * Specifies that the custom view that implements this also places its persisted state into a {@link StateBundle}.
  *
  * This is used by {@link BackstackDelegate#persistViewToState(View)} and {@link BackstackDelegate#restoreViewFromState(View)}.
  */
 public interface Bundleable {
     @NonNull
-    Bundle toBundle();
+    StateBundle toBundle();
 
-    void fromBundle(@Nullable Bundle bundle);
+    void fromBundle(@Nullable StateBundle bundle);
 }