+++ simple-stack/src/test/java/com/zhuinden/simplestack/BackstackDelegateTest.java
@@ -15,7 +15,6 @@
  */
 package com.zhuinden.simplestack;
 
-import org.junit.Ignore;
 import org.junit.Test;
 
 import java.util.ArrayList;
@@ -59,7 +58,6 @@ public class BackstackDelegateTest {
     }
 
     @Test
-    @Ignore // TODO: Bundle.putString is not mocked.
     public void setPersistenceTagAfterOnCreateShouldThrow() {
         BackstackDelegate backstackDelegate = BackstackDelegate.create();
         backstackDelegate.onCreate(null, null, new ArrayList<Object>() {{